pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orbin"
version = "0.1.0"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
description = "Framework for AI-powered chat applications built on FastAPI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
keywords = ["fastapi", "ai", "chat", "framework", "websocket", "sqlalchemy"]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0", 
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "ipython>=8.0.0",
    "jinja2>=3.0.0",
    "inflect>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.800",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/orbin"
"Bug Reports" = "https://github.com/yourusername/orbin/issues"
"Source" = "https://github.com/yourusername/orbin"

[project.scripts]
orbin = "orbin.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["orbin*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
orbin = ["templates/**/*"]

[tool.black]
line-length = 88
target-version = ['py37']

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
